% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List all available palettes}
\usage{
list_palettes()
}
\value{
A vector of character strings of all available palettes.
}
\description{
List the names of all palettes available in technocolour.
}
\examples{
list_palettes()

}
