% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_techno}
\alias{scale_color_techno}
\title{Make a ggplot colour scale}
\usage{
scale_color_techno(palette, n, ...)
}
\arguments{
\item{palette}{The name of the chosen palette. Choices are from: \code{"bodied"}, \code{"esther"},  \code{"hotbot"}, \code{"ibrik"}, \code{"poodle"}, \code{"rush"}, \code{"x"}.}

\item{n}{The integer length of the desired palette. Default value is the length of the specified palette.}
}
\value{
A discrete scale to use for colour in ggplot.
}
\description{
Makes a discrete colour scale of length \code{n} from the chosen technocolour palette.
}
\examples{
library(ggplot2)

ggplot(data = iris,
aes(x = Petal.Length, y = Petal.Width, col = Species)) +
geom_point(size = 3) +
scale_color_techno(palette = "bodied", n = 3)

}
