% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{technocolours}
\alias{technocolours}
\title{Make a technocolour palette}
\usage{
technocolours(name, n = 6, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{rush}, \code{glue}, \code{esther}, \code{hotbot}, \code{ibrik}, \code{poodle}}

\item{n}{The length of the palette, an integer between 1 and 6. Default is n = 6.}
}
\value{
A vector of colours.
}
\description{
Makes a discrete colour palette of length \code{n} from the chosen technocolour palette.
}
\examples{
technocolours(name = "fluorescent_rush", n = 6)

}
