% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{info}
\alias{info}
\title{Print palette information}
\usage{
info(name)
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{flourescent_rush}, \code{glue}, \code{esther}, \code{hot_bot}, \code{ibrik}, \code{poodle_power}}
}
\value{
Prints the track name, artist and URL link for the chosen technocolour palette.
}
\description{
Prints the track name, artist and URL link for the chosen technocolour palette.
}
