% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{info}
\alias{info}
\title{Print record information}
\usage{
info(palette)
}
\arguments{
\item{palette}{The name of the chosen palette. Choices are from: \code{"bodied"}, \code{"esther"},  \code{"hotbot"}, \code{"ibrik"}, \code{"poodle"}, \code{"rush"}, \code{"x"}.}
}
\value{
Prints the track name, artist and URL link for the chosen technocolour palette.
}
\description{
Prints the track name, artist and URL link for the chosen technocolour palette.
}
\examples{
info(palette = "rush")

}
