% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print_palette}
\alias{print_palette}
\title{Print a palette}
\usage{
print_palette(name, n)
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{flourescent_rush}, \code{glue}, \code{esther}, \code{hot_bot}, \code{ibrik}, \code{poodle_power}}

\item{n}{The length of the palette, an integer between 1 and 6.}
}
\description{
Prints the chosen technocolour palette of length \code{n}.
}
\examples{
print_palette(name = "glue", n = 6)

}
