% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_techno}
\alias{scale_color_techno}
\title{Make a ggplot colour scale}
\usage{
scale_color_techno(name, n)
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{flourescent_rush}, \code{glue}, \code{esther}, \code{hot_bot}, \code{ibrik}, \code{poodle_power}}

\item{n}{The length of the palette, an integer between 1 and 6.}
}
\description{
Makes a discrete colour scale of length \code{n} from the chosen technocolour palette.
}
