% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_techno}
\alias{scale_fill_techno}
\title{Make a ggplot fill scale}
\usage{
scale_fill_techno(name, n, ...)
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{"bodied"}, \code{"esther"}, \code{"glue"},  \code{"hotbot"}, \code{"ibrik"}, \code{"poodle"}, \code{"rush"}}

\item{n}{The integer length of the desired palette. Default value is the length of the specified palette.}
}
\value{
A discrete scale to use for fill in ggplot.
}
\description{
Makes a discrete fill scale of length \code{n} from the chosen technocolour palette.
}
\examples{
library(ggplot2)

ggplot(data = iris,
aes(x = Species, y = Petal.Width, fill = Species)) +
geom_violin() +
scale_fill_techno(name = "hotbot", n = 3)

}
