% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{techno_palettes}
\alias{techno_palettes}
\title{Electronic music artwork colour palettes}
\format{
A data frame containing all the colours used in the palette:
\itemize{
  \item palette: The name of the palette
  \item hex: Hex color code
  \item track: The name of the track on which the palette is based
  \item artist: The artist of the track
  \item url: A link to the track
}
}
\usage{
techno_palettes
}
\description{
A dataset containing some colour palettes inspired by electronic music record artwork.
}
\keyword{datasets}
