% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{technocolours}
\alias{technocolours}
\title{Make a technocolour palette}
\usage{
technocolours(name, n)
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{flourescent_rush}, \code{glue}, \code{esther}, \code{hot_bot}, \code{ibrik}, \code{poodle_power}}

\item{n}{The length of the palette, an integer between 1 and 6.}
}
\value{
A vector of colours.
}
\description{
Makes a discrete colour palette of length \code{n} based on the drawn from the chosen technocolour palette.
}
