% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{technocolours}
\alias{technocolours}
\title{Make a technocolour palette}
\usage{
technocolours(name, n, interpolate = TRUE)
}
\arguments{
\item{name}{The name of the chosen palette. Choices are from: \code{"bodied"}, \code{"esther"},  \code{"hotbot"}, \code{"ibrik"}, \code{"poodle"}, \code{"rush"}, \code{"x"}.}

\item{n}{The integer length of the desired palette. Default value is the length of the specified palette.}

\item{interpolate}{Boolean. Indicates whether sequential or diverging palettes should interpolate between colours; this argument is not available for qualitative palettes. Default value is TRUE.}
}
\value{
A vector of colours.
}
\description{
Makes a discrete colour palette of length \code{n} from the chosen technocolour palette.
}
\examples{
technocolours(name = "rush", n = 6)

}
